Renamed src folders to reflect current status
