--- api/src-new/models/Spells/ISpellQuery.ts
@@ -1,8 +1,0 @@
-import IQuery from '../IQuery';
-
-interface ISpellQuery extends IQuery {
-  fp: RegExp | undefined,
-  slot: number | undefined,
-}
-
-export default ISpellQuery;
