+++ api/src-old/controllers/api/spells/dbQueries.ts
