+++ api/src-old/controllers/api/spells/spellResponses/deleteSpell.ts
