+++ api/src-old/controllers/api/spells/spellResponses/getSpell.ts
