+++ api/src-old/controllers/api/spells/spellResponses/getSpells.ts
