+++ api/src-old/models/Spells/ISpellQuery.ts
@@ -1,0 +1,7 @@
+interface ISpellQuery{
+  name: string | RegExp,
+  fp: number,
+  slot: number,
+}
+
+export default ISpellQuery;
