+++ api/src/controllers/error/client/resourceNotFound.ts
