+++ api/src/models/IQuery.ts
