+++ api/src/models/Query.ts
