+++ api/src/models/Spells/ISpellQuery.ts
@@ -1,7 +1,8 @@
-interface ISpellQuery{
-  name: string | RegExp,
-  fp: number,
-  slot: number,
+import IQuery from '../IQuery';
+
+interface ISpellQuery extends IQuery {
+  fp: RegExp | undefined,
+  slot: number | undefined,
 }
 
 export default ISpellQuery;
