+++ api/src/models/Spells/SpellQuery.ts
