+++ api/src/models/Spells/index.ts
