+++ api/src/repositories/mongo/Spells/SpellModel.ts
