+++ api/src/repositories/mongo/Spells/index.ts
